% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec1047-package.R
\docType{package}
\name{ec1047-package}
\alias{ec1047}
\alias{ec1047-package}
\title{ec1047: Code and data for the EC1047 course}
\description{
This package provides code and datasets to be used 
  in the EC1047: Economics of the Public Sector course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcpernias/ec1047}
  \item Report bugs at \url{https://github.com/jcpernias/ec1047/issues}
}

}
\author{
\strong{Maintainer}: José Pernías \email{jcpernias@gmail.com}

}
\keyword{internal}
