% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecv}
\alias{ecv}
\title{Data from the Encuesta de Condiciones de Vida}
\format{A tibble with 143613 observations and 8 variables:
\describe{
  \item{ydisp_hh}{annual disposable income, euros.}
  \item{cu}{consumption units, modified OECD equivalence scale.}
}}
\source{
\href{https://ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=resultados&secc=1254736195153&idp=1254735976608}{Microdata at INE, the Spanish Statistical Office}
}
\usage{
ecv
}
\description{
A small subset of the information contained in the Encuesta de Condiciones
de Vida covering the periood 2008-2018.
}
\keyword{datasets}
