% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Compute Gini coefficient}
\usage{
gini(y, w)
}
\arguments{
\item{y}{numerical vector}

\item{w}{numerical vector with weights for each observation}
}
\value{
a scalar
}
\description{
Compute the Gini coefficient of the distribution
of \code{y}, with weights \code{w}.
}
