% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz.R
\name{lorenz}
\alias{lorenz}
\title{Lorenz curve}
\usage{
lorenz(z, w)
}
\arguments{
\item{z}{a numerical vector.}

\item{w}{a numerical vector with the same length as \code{z}.}
}
\value{
a [tibble][tibble::tibble-package] with the coordinates
of the Lorenz curve.
}
\description{
Compute the coordinates of the Lorenz curve for the distribution
of code \code{z}, using weights \code{w}. If \code{w} is not given,
the unwweighted Lorenz curve is computed.
}
