% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_quantile.R
\name{wtd_quantile}
\alias{wtd_quantile}
\title{Weighted quantiles}
\usage{
wtd_quantile(x, probs, weights = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{probs}{a numerical vector with the quantile probabilities.}

\item{weights}{a numerical vector of weights the same length as \code{x}.}

\item{na.rm}{a logical value indicating wether \code{NA} values in \code{x}
should be stripped before the computation.}
}
\value{
The weighted quantiles of \code{x}
}
\description{
Compute the quantiles of \code{x} weighted by \code{weights}.
}
\details{
If \code{weights} is \code{NULL}, all elements of \code{x} receive the
same weight. Observations with zero weights are omitted from the
computation. Missing values in \code{w} are not handled specially and
produce a missing value as the result.
}
\examples{
wtd_quantile(1:5, weights = 5:1, probs = (1:4)/5)

}
\seealso{
Other weighted statistics: 
\code{\link{wtd_mean}()},
\code{\link{wtd_median}()}
}
\concept{weighted statistics}
