% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Compute Gini coefficient}
\usage{
gini(y, weights = NULL, v = 2)
}
\arguments{
\item{y}{numerical vector.}

\item{weights}{numerical vector of the same length as \code{y}
with weights for each observation.}

\item{v}{numerical scalar, degree of inequality aversion. It
should be non-negative.}
}
\value{
A numeric scalar.
}
\description{
Compute the generalized Gini coefficient of the distribution
of \code{y} with weights \code{weights}. The parameter \code{v}
is the degree of risk aversion. By default, this parameter equals
2 which corresponds with the original Gini coefficient.
}
\details{
Observations with zero weights are discarded before the
computations begin.
}
\examples{
gini(1:5)
gini(1:5, weights = 5:1)
gini(1:5, weights = 5:1, v = 3)

}
