% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_median.R
\name{wtd_median}
\alias{wtd_median}
\title{Weighted median}
\usage{
wtd_median(x, weights = NULL)
}
\arguments{
\item{x}{a numerical vector.}

\item{weights}{a numerical vector of weights the same length as \code{x}.}
}
\value{
The weighted median of \code{x}.
}
\description{
Compute the median of \code{x} weighted by \code{weights}.
}
\details{
If \code{weights} is \code{NULL}, all elements of \code{x} receive the
same weight. Observations with zero weights are omitted from the
computation. Missing values are not handled specially and
produce a missing value as the result.
}
\examples{
wtd_median(1:5, weights = 5:1)

}
\seealso{
Other weighted statistics: 
\code{\link{wtd_mean}()},
\code{\link{wtd_quantile}()}
}
\concept{weighted statistics}
